requirements.txt
# Core app
chainlit==2.4.0
pandas==2.2.3
numpy==1.26.4
openpyxl==3.1.5
xlrd==2.0.1
python-dotenv==1.0.1
openai==1.51.0
pydantic==2.9.2

# LangChain / graph stack
langchain==0.3.2
langchain-core==0.3.9
langchain-community==0.3.1
langchain-text-splitters==0.3.0
langchain-openai==0.2.2
langchain-anthropic==0.2.0
langchain-experimental==0.3.2
langgraph==0.2.34
langgraph-checkpoint==2.0.1
tiktoken==0.8.0
httpx-sse==0.4.0

# Utility packages used directly in code/tests
tenacity==8.5.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
yarl==1.13.1
networkx==3.3
pulp==2.9.0
pytest==8.3.3
colorama==0.4.6